Add header and side bar to layout
