+++ src/app/layout.tsx
@@ -6,6 +6,7 @@ import {
 } from 'next/font/google'
 import './globals.css'
 import FooterBar from '@/components/FooterBar'
+import Link from 'next/link'
 
 const playfair = Playfair_Display({
   variable: '--font-heading',
@@ -76,6 +77,49 @@ export const metadata: Metadata = {
   }
 }
 
+// Document categories - easily extensible for future document types
+const documentCategories = [
+  {
+    id: 'technical',
+    name: 'Technical Documents',
+    description:
+      'Explore detailed documents, cultural descriptions, and scholarly guides that illuminate how this cassowary civilization functions, thrives, and shapes its remarkable world.',
+    href: '/technical-docs',
+    icon: '🔧'
+  },
+  {
+    id: 'speeches',
+    name: 'Speeches',
+    description:
+      'Listen to the voices of cassowary leaders, orators, and visionaries as they address their society with wisdom, inspiration, and calls to action.',
+    href: '/speeches',
+    icon: '🎤'
+  },
+  {
+    id: 'concept-art',
+    name: 'Concept Art',
+    description:
+      'Explore visual designs, sketches, and artistic renderings that bring the cassowary world to life through the eyes of its creators.',
+    href: '/concept-art',
+    icon: '🎨'
+  }
+  // Future categories can be added here:
+  // {
+  //   id: 'research',
+  //   name: 'Research Papers',
+  //   description: 'Academic research and scientific papers',
+  //   href: '/research',
+  //   icon: '📚'
+  // },
+  // {
+  //   id: 'user-guides',
+  //   name: 'User Guides',
+  //   description: 'User manuals and tutorials',
+  //   href: '/user-guides',
+  //   icon: '📖'
+  // }
+]
+
 export default function RootLayout({
   children
 }: Readonly<{
@@ -84,9 +128,65 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`${playfair.variable} ${crimson.variable} ${mono.variable} antialiased  flex flex-col`}
+        className={`${playfair.variable} ${crimson.variable} ${mono.variable} antialiased flex flex-col`}
+        style={{ backgroundColor: 'var(--background)' }}
       >
-        <div className="flex-1">{children}</div>
+        {/* Header */}
+        <header
+          className="border-b border-solid"
+          style={{ borderColor: 'var(--color-leaf-shadow)' }}
+        >
+          <div className="p-6">
+            <Link href="/" className="inline-block">
+              <h1
+                className="text-4xl font-bold hover:opacity-80 transition-opacity duration-200"
+                style={{ color: 'var(--color-cassowary)' }}
+              >
+                Cassowary World
+              </h1>
+            </Link>
+          </div>
+        </header>
+
+        {/* Main content area with sidebar */}
+        <div className="flex flex-1">
+          {/* Left Sidebar */}
+          <aside className="sidebar">
+            <div className="p-6">
+              <h2
+                className="text-2xl font-bold mb-6"
+                style={{ color: 'var(--color-cassowary)' }}
+              >
+                Document Categories
+              </h2>
+              <nav className="space-y-3">
+                {documentCategories.map(category => (
+                  <Link
+                    key={category.id}
+                    href={category.href}
+                    className="category-card"
+                  >
+                    <div className="flex items-center space-x-3">
+                      <span className="text-2xl">{category.icon}</span>
+                      <div>
+                        <h3 className="font-semibold text-lg">
+                          {category.name}
+                        </h3>
+                        <p className="text-sm opacity-90">
+                          {category.description}
+                        </p>
+                      </div>
+                    </div>
+                  </Link>
+                ))}
+              </nav>
+            </div>
+          </aside>
+
+          {/* Main Content */}
+          <main className="flex-1">{children}</main>
+        </div>
+
         <FooterBar />
       </body>
     </html>
