+++ src/app/speeches/[slug]/page.tsx
@@ -33,7 +33,7 @@ export default async function SpeechPage({
     : null
 
   return (
-    <main className="container mx-auto  p-8 flex flex-col gap-4">
+    <div className="p-8 flex flex-col gap-4">
       <Link href="/speeches" className="back-link">
         ← Back to speeches
       </Link>
@@ -61,6 +61,6 @@ export default async function SpeechPage({
           {doc.markdown}
         </ReactMarkdown>
       </article>
-    </main>
+    </div>
   )
 }
